backend/requirements.txt
fastapi[standard]>=0.116.0
uvicorn[standard]>=0.35.0
python-dotenv>=1.1.1
stripe>=12.2.0
firebase-admin>=6.9.0
google-generativeai>=0.8.5
python-multipart>=0.0.20
pydantic>=2.11.6
openpyxl>=3.1.2
pandas>=2.0.0
google-cloud-storage>=2.10.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
arq>=0.26.0
redis>=5.0.0
